refactor: entity에서 도메인 제거
