+++ app/src/main/java/woowacourse/movie/data/MovieConverter.kt
@@ -1,8 +1,6 @@
 package woowacourse.movie.data
 
 import androidx.room.TypeConverter
-import woowacourse.movie.mapper.toUiModel
-import woowacourse.movie.model.Seat
 import woowacourse.movie.util.Formatter.dateHyphenFormatter
 import woowacourse.movie.util.Formatter.timeFormatter
 import java.time.LocalDate
@@ -20,21 +18,4 @@ class MovieConverter {
 
     @TypeConverter
     fun toTime(time: String): LocalTime = LocalTime.parse(time, timeFormatter)
-
-    @TypeConverter
-    fun fromSeats(seats: Set<Seat>): String {
-        return seats.map { it.toUiModel() }.joinToString(", ") { point ->
-            "${'A' + point.row}${point.col + 1}"
-        }
-    }
-
-    @TypeConverter
-    fun toSeats(seats: String): Set<Seat> {
-        return seats.split(", ")
-            .map { seat ->
-                val row = seat[0] - 'A'
-                val col = seat.substring(1).toInt() - 1
-                Seat(row, col)
-            }.toSet()
-    }
 }
