+++ app/src/main/java/woowacourse/movie/data/entity/TicketEntity.kt
@@ -3,7 +3,6 @@ package woowacourse.movie.data.entity
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import woowacourse.movie.model.Seat
 import java.time.LocalDate
 import java.time.LocalTime
 
@@ -15,7 +14,7 @@ data class TicketEntity(
     @ColumnInfo(name = "date") val date: LocalDate,
     @ColumnInfo(name = "time") val time: LocalTime,
     @ColumnInfo(name = "head_count") val headCount: Int,
-    @ColumnInfo(name = "seat") val seat: Set<Seat>,
+    @ColumnInfo(name = "seat") val seat: String,
     @ColumnInfo(name = "theater") val theater: String,
     @ColumnInfo(name = "price") val price: Int,
 )
