+++ app/src/main/java/woowacourse/movie/mapper/TicketEntityMapper.kt
@@ -2,6 +2,7 @@ package woowacourse.movie.mapper
 
 import woowacourse.movie.data.entity.TicketEntity
 import woowacourse.movie.model.HeadCount
+import woowacourse.movie.model.Seat
 import woowacourse.movie.model.Seats
 import woowacourse.movie.model.Ticket
 
@@ -11,7 +12,7 @@ fun Ticket.toEntity(): TicketEntity {
         date = selectedDate,
         time = selectedTime,
         headCount = headCount.value,
-        seat = seats.seats,
+        seat = fromSeats(seats),
         theater = theater,
         price = amount,
     )
@@ -24,6 +25,23 @@ fun TicketEntity.toDomain(): Ticket {
         headCount = HeadCount(headCount),
         selectedDate = date,
         selectedTime = time,
-        seats = Seats(seat),
+        seats = toSeats(seat),
+    )
+}
+
+private fun fromSeats(seats: Seats): String {
+    return seats.seats.map { it.toUiModel() }.joinToString(", ") { point ->
+        "${'A' + point.row}${point.col + 1}"
+    }
+}
+
+private fun toSeats(seats: String): Seats {
+    return Seats(
+        seats.split(", ")
+            .map { seat ->
+                val row = seat[0] - 'A'
+                val col = seat.substring(1).toInt() - 1
+                Seat(row, col)
+            }.toSet(),
     )
 }
